requirements.txt
streamlit
numpy
opencv-python-headless
Pillow
pandas
deepface